+++ example/example.h
@@ -9,7 +9,7 @@
 #define EXAMPLE_H_
 
 /* example.c */
-void	read_data (char *fn, int skip_header, size_t *n, size_t *p, double **y, double **x);
+void	read_data (char *fn, int skip_header, int *n, int *p, double **y, double **x);
 
 /* example_larsen.c */
 void	example_larsen (const linreg *lreg, double start, double dt, double stop, double gamma, int maxiter);
